src/data-service/requirements.txt
# Data Service dependencies
fastapi>=0.116,<1.0
uvicorn>=0.30,<1.0
pydantic[email]>=2.8,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.4,<3.0
email-validator>=2.1,<3.0
python-multipart>=0.0.9,<1.0
pdfplumber>=0.10,<1.0
chardet>=5.0,<6.0
